+++ src/main/kotlin/cloud/glitchdev/rfu/events/managers/ContainerEvents.kt
@@ -4,7 +4,7 @@ import cloud.glitchdev.rfu.events.AbstractEventManager
 import net.minecraft.world.item.ItemStack
 
 object ContainerEvents : AbstractEventManager<(containerId : Int, itens : List<ItemStack>) -> Unit, ContainerEvents.ContainerOpenEvent>() {
-    fun runTasks(containerId : Int, itens : List<ItemStack>) {
+    override val runTasks: (Int, List<ItemStack>) -> Unit = { containerId, itens ->
         safeExecution {
             tasks.forEach { event -> event.callback(containerId, itens) }
         }