+++ src/main/kotlin/cloud/glitchdev/rfu/events/managers/EntityRenderEvents.kt
@@ -5,7 +5,7 @@ import cloud.glitchdev.rfu.events.wrappers.Cancelable
 import net.minecraft.world.entity.Entity
 
 object EntityRenderEvents : AbstractEventManager<(entity : Entity, event : Cancelable<Boolean>) -> Unit, EntityRenderEvents.EntityRenderEvent>() {
-    fun runTasks(entity : Entity, event : Cancelable<Boolean>) {
+    override val runTasks: (Entity, Cancelable<Boolean>) -> Unit = { entity, event ->
         safeExecution {
             tasks.forEach { task -> task.callback(entity, event) }
         }