+++ src/main/kotlin/cloud/glitchdev/rfu/events/managers/SetSlotEvents.kt
@@ -1,11 +1,10 @@
 package cloud.glitchdev.rfu.events.managers
 
-import cloud.glitchdev.rfu.RiccioFishingUtils.mc
 import cloud.glitchdev.rfu.events.AbstractEventManager
 import net.minecraft.world.item.ItemStack
 
 object SetSlotEvents : AbstractEventManager<(containerId: Int, slot: Int, item: ItemStack) -> Unit, SetSlotEvents.SetSlotEvent>() {
-    fun runTasks(containerId: Int, slot: Int, item: ItemStack) {
+    override val runTasks: (Int, Int, ItemStack) -> Unit = { containerId, slot, item ->
         safeExecution {
             tasks.forEach { event -> event.callback(containerId, slot, item) }
         }